org restrictions in forms and controllers

+++ app/controllers/organizations_controller.rb
@@ -4,21 +4,27 @@ class OrganizationsController < ApplicationController
   layout 'organization'
 
   def index
+    redirect_to root_url unless current_user.has_role? :admin
     @organizations = Organization.all
     #@new_organization = Organization.new
   end
 
   def show
     @organization = Organization.find(params[:id])
+    unless (current_user.has_role? :admin || current_user.owned_organizations.include?(@organization) || current_user.joined_organizations.include?(@organization))
+      redirect_to root_url
+    end
     not_found unless @organization.present?
   end
 
   def edit
+    redirect_to root_url unless current_user.has_role? :admin
     @organization = Organization.find(params[:id])
     not_found unless @organization.present?
   end
 
   def new
+    redirect_to root_url unless current_user.has_role? :admin
     if current_user.has_role? :admin
       @organization = Organization.new
     else
@@ -28,6 +34,7 @@ class OrganizationsController < ApplicationController
   end
 
   def create
+    return current_user.has_role? :admin
     @organization = Organization.new(permitted_params)
     if @organization.save
       flash[:notice] = "#{t('activerecord.models.organization', count: 1)} добавлен"
@@ -38,6 +45,7 @@ class OrganizationsController < ApplicationController
   end
 
   def update
+    return unless current_user.has_role? :admin
     @organization = Organization.find(params[:id])
     if @organization.update_attributes(permitted_params)
       flash[:notice] = "#{t('activerecord.models.organization', count: 1)} обновлен"
@@ -48,6 +56,7 @@ class OrganizationsController < ApplicationController
   end
 
   def destroy
+    return unless current_user.has_role? :admin
     @organization = Organization.find(params[:id])
     if @organization.destroy
       flash[:notice] = "#{t('activerecord.models.organization', count: 1)} удален"
